Updated player & enemy collision boxes
Player collision shapes have been updated to not include the tail. The center of the player is where the player will take damage so that the user doesnt feel that it is unfair.

Enemy collisions changed so that the enemy can do damage to the player if they are directly above or below.

Player wall collision also changed so that they slide against the wall instead of getting caught on corners.

Changed Player damage_player() function and added invincibility frames.